Refactor Ms.Ext.Azure.LogForwarder slightly
