+++ sdk/extensions/Microsoft.Extensions.Azure/src/AzureEventSourceLogForwarder.cs
@@ -2,9 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Collections;
 using System.Collections.Concurrent;
-using System.Collections.Generic;
 using System.Diagnostics.Tracing;
 using Azure.Core.Diagnostics;
 using Azure.Core.Shared;
@@ -88,33 +86,6 @@ namespace Microsoft.Extensions.Azure
         private static string FormatMessage(EventSourceEvent eventSourceEvent, Exception exception)
         {
             return EventSourceEventFormatting.Format(eventSourceEvent.EventData);
-        }
-
-        private readonly struct EventSourceEvent: IReadOnlyList<KeyValuePair<string, object>>
-        {
-            public EventWrittenEventArgs EventData { get; }
-
-            public EventSourceEvent(EventWrittenEventArgs eventData)
-            {
-                EventData = eventData;
-            }
-
-            public IEnumerator<KeyValuePair<string, object>> GetEnumerator()
-            {
-                for (int i = 0; i < Count; i++)
-                {
-                    yield return new KeyValuePair<string, object>(EventData.PayloadNames[i], EventData.Payload[i]);
-                }
-            }
-
-            IEnumerator IEnumerable.GetEnumerator()
-            {
-                return GetEnumerator();
-            }
-
-            public int Count => EventData.PayloadNames.Count;
-
-            public KeyValuePair<string, object> this[int index] => new KeyValuePair<string, object>(EventData.PayloadNames[index], EventData.Payload[index]);
         }
     }
 }
